array iterator

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -13,7 +13,7 @@ void array_iterator(int *array, size_t size, void (*action)(int))
 {
 	if (array != NULL && action != NULL)
 	{
-		for (unsigned int i = 0; i < size; i++)
+		for (size_t i = 0; i < size; i++)
 			action(array[i]);
 	}
 }
